Added extra tests to check for arithmetic part and pretty printing. Changed eval to return double instead of int.

+++ a1-extend-tiny-evaluator/tiny/eval.h
@@ -3,6 +3,6 @@
 
 #include "tree.h"
 
-int evalEXP(EXP *e);
+double evalEXP(EXP *e);
 
 #endif /* !EVAL_H */
